requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

# blueapps requirement

redis==5.0.3
Django==3.2.25
celery==5.4.0
eventlet==0.33.3
greenlet==3.0.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
django-filter==23.5

# jwt
pyCryptodome==3.23.0
PyJWT==1.7.1
cryptography==43.0.3

mysqlclient==2.1.1
MarkupSafe==1.1.1
Mako==1.1.4
requests==2.28.1
python-json-logger==2.0.1
whitenoise==5.2.0
gunicorn==23.0.0

blueapps[opentelemetry]==4.15.6
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-instrumentation-dbapi==0.42b0
opentelemetry-instrumentation-django==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-wsgi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
prometheus-client==0.24.1
pymysql==1.1.1
pre-commit==4.5.1

setuptools==58.0.1